chore: reuse CategoryOptionMapper
